requirements.txt
flask==2.3.3
flask-sqlalchemy==3.0.5
flask-login==0.6.2
flask-migrate==4.0.4
python-dotenv==1.0.0
flask-wtf==1.1.1
email-validator==2.0.0
werkzeug==2.3.7
flask-mail==0.9.1
